requirements.txt
boto3>=1.34.0
botocore>=1.34.0
pyyaml>=6.0
rich>=13.0.0
